script/requirements.txt
xgboost==1.3.3
cmapy==0.6.6
opencv_python==4.5.1.48
numpy==1.19.2
torch==1.9.0
matplotlib==3.3.2
torchvision==0.2.1
tqdm==4.50.2
networks==0.3.7
Pillow==8.2.0
